+++ apps/greyhound_markdown/client/lib/src/application/application.dart
@@ -1,2 +1,3 @@
 export 'room/room_id.dart';
+export 'user_settings/app_language.dart';
 export 'user_settings/user_settings_cubit.dart';